requirements.txt
apt-xapian-index==0.49
autopep8==1.5.5
backcall==0.2.0
blinker==1.4
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
google-api-core==1.25.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-vision==2.1.0
googleapis-common-protos==1.52.0
gpg===1.13.1-unknown
grpcio==1.35.0
httplib2==0.14.0
idna==2.8
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libcst==0.3.16
Markdown==3.1.1
matplotlib==3.1.2
matplotlib-venn==0.11.5
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.20.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
packaging==20.3
panda==0.3.1
parso==0.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.16
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.7
scipy==1.3.3
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
systemd-python==234
toml==0.10.2
tornado==6.1
traitlets==5.0.5
ttystatus==0.38
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.2.5
xkit==0.0.0
